+++ src/main/java/gift/product/application/ProductService.java
@@ -98,71 +98,6 @@ public class ProductService {
                 .toList();
     }
 
-    public class ProductResponse {
-        private Long id;
-        private String name;
-        private Double price;
-        private String imageUrl;
-
-        public ProductResponse(Long id, String name, Double price, String imageUrl) {
-            this.id = id;
-            this.name = name;
-            this.price = price;
-            this.imageUrl = imageUrl;
-        }
-
-        public Long getId() {
-            return id;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public Double getPrice() {
-            return price;
-        }
-
-        public String getImageUrl() {
-            return imageUrl;
-        }
-    }
-
-    public class ProductListResponse {
-        private List<ProductResponse> content;
-        private int number;
-        private int totalElement;
-        private int size;
-        private boolean last;
-
-        public ProductListResponse(List<ProductResponse> content, int number, int totalElement, int size, boolean last) {
-            this.content = content;
-            this.number = number;
-            this.totalElement = totalElement;
-            this.size = size;
-            this.last = last;
-        }
-
-        public List<ProductResponse> getContent() {
-            return content;
-        }
-
-        public int getNumber() {
-            return number;
-        }
-
-        public int getTotalElement() {
-            return totalElement;
-        }
-
-        public int getSize() {
-            return size;
-        }
-
-        public boolean isLast() {
-            return last;
-        }
-    }
 
     public class ProductOptionData {
         private Long id;
