+++ src/main/java/gift/product/application/WishListService.java
@@ -10,12 +10,12 @@ import gift.util.ErrorCode;
 import jakarta.transaction.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Objects;
+import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -32,16 +32,22 @@ public class WishListService {
         this.userService = userService;
     }
 
-    public WishList getWishListByUserId(Long userId) {
-        return wishListRepository.findByUserId(userId);
+
+    public List<WishListResponse> getWishListByUserId(Long userId, Pageable pageable) {
+        if (wishListRepository.findByUserId(userId, pageable).isEmpty()) {
+            wishListRepository.save(new WishList(userService.getUser(userId), LocalDateTime.now()));
+        }
+        Page<WishList> wishLists = wishListRepository.findByUserId(userId, pageable);
+
+        return wishLists.stream()
+                .map(wishList -> new WishListResponse(wishList.getId(),
+                        wishList.getWishListProducts().stream()
+                                .map(WishListProduct::getProduct)
+                                .map(ProductResponse::new)
+                                .collect(Collectors.toList())))
+                .collect(Collectors.toList());
     }
 
-    public Page<WishList> getProductsInWishList(Long userId, Pageable pageable) {
-        if (wishListRepository.findByUserId(userId, pageable).isEmpty()) {
-            throw new ProductException(ErrorCode.WISHLIST_NOT_FOUND);
-        }
-        return wishListRepository.findByUserId(userId, pageable);
-    }
 
     @Transactional
     public void addProductToWishList(AddWishListRequest request) {
