+++ src/main/java/gift/product/domain/ProductResponse.java
@@ -1,0 +1,41 @@
+package gift.product.domain;
+
+public class ProductResponse {
+    private Long id;
+    private String name;
+    private Double price;
+    private String imageUrl;
+
+    public ProductResponse(Long id, String name, Double price, String imageUrl) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.imageUrl = imageUrl;
+    }
+
+    public ProductResponse() {
+    }
+
+    public ProductResponse(Product product) {
+        this.id = product.getId();
+        this.name = product.getName();
+        this.price = product.getPrice();
+        this.imageUrl = product.getImageUrl();
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Double getPrice() {
+        return price;
+    }
+
+    public String getImageUrl() {
+        return imageUrl;
+    }
+}
